dm Cache Functions

Functions for reading, writing, and checking/checksumming caches of dm objects

Methods

Public methods

DMCache$write()

DMCache$read()

DMCache$check()

DMCache$checksum()

Method write()

Usage

DMCache$write(
  dm,
  checksum,
  file,
  dir = path_create("~", ".cache_dmhct"),
  n_threads = max(1L, parallel::detectCores()%/%2L)
)

Arguments

dm

[dm] A dm object

checksum

[chr(1)] The hash of the dm object

file

[chr(1)] The file stem to use for the cache objects

dir

[chr(1)] The directory in which to save the cache objects

n_threads

[int(1)] The number of threads to use when serializing the dm object

Returns

[dm] The dm object (invisibly) Read dm Data or its Checksum from a the Specified Cache

Method read()

Usage

DMCache$read(
  obj = c("data", "checksum"),
  file,
  dir = path_create("~", ".cache_dmhct"),
  fail = TRUE,
  n_threads = if (obj == "data") max(1L, parallel::detectCores()%/%2L) else 1L
)

Arguments

obj

[chr(1)] What to read from the cache

file

[chr(1)] The file stem to use for the cache objects

dir

[chr(1)] The directory in which to save the cache objects

fail

[lgl(1)] Whether to error if dir/obj_file combo does not exist

n_threads

[int(1)] The number of threads to use when de-serializing the object

Returns

[dm|chr(1)] The de-serialized object Check Hash for dm Object

Method check()

Usage

DMCache$check(
  checksum,
  file,
  dir = path_create("~", ".cache_dmhct"),
  n_threads = max(1L, parallel::detectCores()%/%2L)
)

Arguments

checksum

[chr(1)] The hash of the dm object

file

[chr(1)] The file stem to use for the cache objects

dir

[chr(1)] The directory in which to save the cache objects

n_threads

int(1) The number of threads to use when de-serializing the cached checksum

Returns

[lgl(1)] Whether the passed checksum and cached checksum are identical Hash a dm Object

dm_checksum() creates a checksum by hashing each table + its name individually, sorting by name, and hashing a list of the results. This means that the hash does not change when the ordering of the tables change in the dm, but does change if any data or names are added, deleted, or modified.

Method checksum()

Usage

DMCache$checksum(dm)

Arguments

dm

[dm] The dm object to hash

Returns

[chr(1)] The hash string of the object