Utility Functions for Chimerism Table ELT

Collection of utility functions for chimerism data

Methods

Public methods

UtilsChimerism$str_detect_fct()

UtilsChimerism$str_range_to_num()

UtilsChimerism$str_excel_dt_to_num()

UtilsChimerism$normalize()

UtilsChimerism$probit()

UtilsChimerism$inv_probit()

Method str_detect_fct()

Usage

UtilsChimerism$str_detect_fct(x, lvl, lbl)

Arguments

x

[chr] A character vector

lvl

[chr(1)] The level's numeric representation

lbl

[chr(1)] The level's label

Returns

[lgl(1)] Presence or absence of the indicated factor level Convert a % Range to It's Midpoint

Method str_range_to_num()

Usage

UtilsChimerism$str_range_to_num(x)

Arguments

x

[chr] Character representation of a range of percentages

Returns

[dbl] The midpoint of each range of values Convert Excel Date to Number

Converts Excel dates in MDY format to a number using an origin of "1/0/1900". Elements that cannot be converted are returned as-is.

Method str_excel_dt_to_num()

Usage

UtilsChimerism$str_excel_dt_to_num(x)

Arguments

x

[chr] Character vector containing Excel dates

Returns

[chr] Numeric representation of Excel date, where available Normalize a numeric Vector to [0, 1]

Method normalize()

Usage

UtilsChimerism$normalize(x)

Arguments

x

[dbl] A numeric vector or a vector than can be coerced to numeric

Returns

[dbl] The normalized vector Transform a numeric Vector Using the Probit Function

Method probit()

Transforms a numeric vector using the Gaussian quantile function qnorm(). Converts the vector to the [0, 1] range using range prior to transformation. Since qnorm(0) and qnorm(1) are infinite, values closer than tol to 0 or 1 are truncated before transformation.

Usage

UtilsChimerism$probit(x, range = c(0, 100), tol = sqrt(.Machine$double.eps))

Arguments

x

[dbl] A vector to transform

range

[dbl(2)] The lower and upper bounds of the domain of x

tol

[dbl(1)] The tolerance for truncating x on the [0, 1] scale; must be between 0 and 0.01.

Returns

[dbl] The transformed vector Transform a numeric Vector Using the Inverse Probit Function

Method inv_probit()

Transforms a numeric vector using the Gaussian distribution function pnorm(). As the inverse of probit(), also inverts truncation by replacing transformed values closer than tol to 0 or 1 with 0 or 1 (respectively). Maps result to the domain specified by range; values outside of range will be truncated.

Usage

UtilsChimerism$inv_probit(
  x,
  range = c(0, 100),
  tol = sqrt(.Machine$double.eps)
)

Arguments

x

[dbl] A vector to transform

range

[dbl(2)] The lower and upper bounds of the range of the result

tol

[dbl(1)] The tolerance used for truncating in the probit transform

Returns

[dbl] The transformed vector