Extract, Load, and Transform Remote Tables to Local Source

dm_elt() encompasses the entire legacy dmhct pipeline; however, this pipeline is deprecated no longer under active development. While this function will be retained for backwards compatibility, it is strongly recommended that new code use the new pipeline instead.

Usage

dm_elt(dm_remote = dm_sql_server(), reset = FALSE, close = NULL)

Arguments

dm_remote

[dm] Remote dm object containing HCT data

reset

[lgl(1)] Should the cache be reset to the current results, even if inputs have not changed? This is useful if data processing logic has changed, but the underlying data have not.

close

[lgl(1)] Whether to close the SQL Server connection on exit. NULL closes if dm_remote has attribute default == TRUE and leaves open otherwise.

Value

[dm] A dm object