Select and Convert Table + Columns From Remote Source

dm_extract() selects tables and columns of potential interest. It combines all Cerner tables into one and joins HLA tables.

Usage

dm_extract(
  dm_remote = dm_sql_server(),
  collect = TRUE,
  compute = FALSE,
  cache = collect,
  reset = FALSE
)

Arguments

dm_remote

[dm] A dm object connected to the SQL Server for MLinHCT

collect

[lgl(1)] Should tables be collected locally on output? collect = TRUE means that tables will also be computed, even if compute = FALSE.

compute

[lgl(1)] Should tables be computed on the server? This does not mean that the computed tables will be transferred to the local machine; to do so, set collect = TRUE.

cache

[lgl(1)] Should results be cached (if inputs have changed) or read from cache (if inputs have not changed)?

reset

[lgl(1)] Should the cache be reset to the current results, even if inputs have not changed? This is useful if data processing logic has changed, but the underlying data have not.

Value

[dm] The updated dm object